+++ src/api_keys.rs
@@ -83,6 +83,7 @@ pub mod types {
     impl Deref for ApiKeyId {
         type Target = str;
 
+        #[inline]
         fn deref(&self) -> &Self::Target {
             self.as_ref()
         }